requirements.txt
astropy==3.1.2
cycler==0.10.0
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
emcee==2.2.1
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
PyAstronomy==0.13.0
PyJWT==1.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
quantities==0.12.3
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
